Rename assignment.js to def.js
