+++ src/functions.js
@@ -10,10 +10,10 @@ export default functions;
 import "./functions/accent";
 import "./functions/accentunder";
 import "./functions/arrow";
-import "./functions/assignment";
 import "./functions/char";
 import "./functions/color";
 import "./functions/cr";
+import "./functions/def";
 import "./functions/delimsizing";
 import "./functions/enclose";
 import "./functions/environment";
